Extract PersonaCoin and PersonaPresence from Persona
